feat(KEN-706): settings edits compose into the plan [no-changelog]
A manifest save re-plans the scope, and that plan may seed or refresh
kendex.settings.toml itself, so a settings edit cannot be a second write:
its precondition would bind to bytes the first write had already replaced.
Edits are an input to the plan instead. Seeds, comment refreshes and a
person's values become one WriteFile under one precondition, with the
ledger they all move riding out on the one lock the pass writes — so a key
seeded and set in the same save lands once, and a later comment refresh
still recognises the block as seeding's own.

Reading is the other half. Per installed skill, one of four states —
ships no template, nothing here could read one, the strict reader refused
it, or rows — because a template the strict reader refuses may still have
seeded keys through the lenient reader, and calling that "nothing" would
misreport a working file. Each row's current value is three-state for the
same reason: only a single readable [env] assignment is comparable with a
default, and a key assigned twice or assigned where no script reads it
says so with its lines rather than showing a value nobody gets. Global
scope answers known-empty rather than leaving the question open.

An edit moves exactly the key's quoted span, so comments, trailing notes
and CRLF terminators survive; values are held to the loaders' grammar; a
key nothing can read refuses naming the lines; and the writer repeats
seeding's refusal to write through anything but a regular file.

The refusals live beside the code that raises them rather than swelling
the crate's error enum: one transparent variant carries the four, each
with the key, the file or the lines a person acts on.

No changelog entry: nothing here is reachable from the app yet. KEN-707
builds the Customize surface on it and carries the entry for both.

+++ crates/core/src/settings_template/tests.rs
@@ -4,7 +4,7 @@ const GOOD: &str = "[env]\n\n# What the reviewers do.\n# Comma separated.\nREVIE
 
 /// Every finding's problem text, so a fixture asserts what was said and
 /// where without repeating the whole struct.
-fn located(text: &str) -> Vec<(usize, String)> {
+fn located(text: &str) -> Vec<(u32, String)> {
     read(text)
         .findings
         .into_iter()
